Class Diagram
A UML of a Riyadh Train Integrated System project. 
The Riyadh Metro Integrated System is an advanced software platform developed to monitor all
the aspects of Riyadh Metro operations, such as station oversight, route optimization, Online
ticketing, and fare management. This system aims to enhance the quality of the customer
experience, and increase the overall efficiency of the operation.